+++ android/sdk/src/main/java/org/jitsi/meet/sdk/JitsiMeetView.java
@@ -16,36 +16,33 @@
 
 package org.jitsi.meet.sdk;
 
+import android.app.Activity;
 import android.content.Context;
 import android.os.Bundle;
 import android.util.AttributeSet;
+import android.widget.FrameLayout;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.facebook.react.bridge.ReadableMap;
+import com.facebook.react.ReactRootView;
+import com.rnimmersive.RNImmersiveModule;
 
 import org.jitsi.meet.sdk.log.JitsiMeetLogger;
 
-import java.lang.reflect.Method;
-import java.util.Map;
 
-public class JitsiMeetView extends BaseReactView<JitsiMeetViewListener>
-        implements OngoingConferenceTracker.OngoingConferenceListener {
+public class JitsiMeetView extends FrameLayout {
 
     /**
-     * The {@code Method}s of {@code JitsiMeetViewListener} by event name i.e.
-     * redux action types.
+     * Background color used by {@code BaseReactView} and the React Native root
+     * view.
      */
-    private static final Map<String, Method> LISTENER_METHODS
-        = ListenerUtils.mapListenerMethods(JitsiMeetViewListener.class);
+    private static int BACKGROUND_COLOR = 0xFF111111;
 
     /**
-     * The URL of the current conference.
+     * React Native root view.
      */
-    // XXX Currently, one thread writes and one thread reads, so it should be
-    // fine to have this field volatile without additional synchronization.
-    private volatile String url;
+    private ReactRootView reactRootView;
 
     /**
      * Helper method to recursively merge 2 {@link Bundle} objects representing React Native props.
@@ -109,10 +106,19 @@ public class JitsiMeetView extends BaseReactView<JitsiMeetViewListener>
         initialize(context);
     }
 
-    @Override
+    /**
+     * Releases the React resources (specifically the {@link ReactRootView})
+     * associated with this view.
+     *
+     * MUST be called when the {@link Activity} holding this view is destroyed,
+     * typically in the {@code onDestroy} method.
+     */
     public void dispose() {
-        OngoingConferenceTracker.getInstance().removeListener(this);
-        super.dispose();
+        if (reactRootView != null) {
+            removeView(reactRootView);
+            reactRootView.unmountReactApplication();
+            reactRootView = null;
+        }
     }
 
     /**
@@ -130,8 +136,7 @@ public class JitsiMeetView extends BaseReactView<JitsiMeetViewListener>
                 PictureInPictureModule.class);
         if (pipModule != null
                 && pipModule.isPictureInPictureSupported()
-                && !JitsiMeetActivityDelegate.arePermissionsBeingRequested()
-                && this.url != null) {
+                && !JitsiMeetActivityDelegate.arePermissionsBeingRequested()) {
             try {
                 pipModule.enterPictureInPicture();
             } catch (RuntimeException re) {
@@ -151,10 +156,40 @@ public class JitsiMeetView extends BaseReactView<JitsiMeetViewListener>
     }
 
     /**
-     * Leaves the currently active conference.
+     * Creates the {@code ReactRootView} for the given app name with the given
+     * props. Once created it's set as the view of this {@code FrameLayout}.
+     *
+     * @param appName - The name of the "app" (in React Native terms) to load.
+     * @param props - The React Component props to pass to the app.
      */
-    public void leave() {
-        setProps(new Bundle());
+    private void createReactRootView(String appName, @Nullable Bundle props) {
+        if (props == null) {
+            props = new Bundle();
+        }
+
+        if (reactRootView == null) {
+            reactRootView = new ReactRootView(getContext());
+            reactRootView.startReactApplication(
+                ReactInstanceManagerHolder.getReactInstanceManager(),
+                appName,
+                props);
+            reactRootView.setBackgroundColor(BACKGROUND_COLOR);
+            addView(reactRootView);
+        } else {
+            reactRootView.setAppProperties(props);
+        }
+    }
+
+    private void initialize(@NonNull Context context) {
+        // Check if the parent Activity implements JitsiMeetActivityInterface,
+        // otherwise things may go wrong.
+        if (!(context instanceof JitsiMeetActivityInterface)) {
+            throw new RuntimeException("Enclosing Activity must implement JitsiMeetActivityInterface");
+        }
+
+        setBackgroundColor(BACKGROUND_COLOR);
+
+        ReactInstanceManagerHolder.initReactInstanceManager((Activity) context);
     }
 
     /**
@@ -179,46 +214,27 @@ public class JitsiMeetView extends BaseReactView<JitsiMeetViewListener>
         createReactRootView("App", props);
     }
 
-    /**
-     * Handler for {@link OngoingConferenceTracker} events.
-     * @param conferenceUrl
-     */
-    @Override
-    public void onCurrentConferenceChanged(String conferenceUrl) {
-        // This property was introduced in order to address
-        // an exception in the Picture-in-Picture functionality which arose
-        // because of delays related to bridging between JavaScript and Java. To
-        // reduce these delays do not wait for the call to be transferred to the
-        // UI thread.
-        this.url = conferenceUrl;
-    }
-
-    /**
-     * Handler for {@link ExternalAPIModule} events.
-     *
-     * @param name The name of the event.
-     * @param data The details/specifics of the event to send determined
-     * by/associated with the specified {@code name}.
-     */
-    @Override
-    @Deprecated
-    protected void onExternalAPIEvent(String name, ReadableMap data) {
-        onExternalAPIEvent(LISTENER_METHODS, name, data);
-    }
-
     @Override
     protected void onDetachedFromWindow() {
         dispose();
         super.onDetachedFromWindow();
     }
 
-    private void initialize(@NonNull Context context) {
-        // Check if the parent Activity implements JitsiMeetActivityInterface,
-        // otherwise things may go wrong.
-        if (!(context instanceof JitsiMeetActivityInterface)) {
-            throw new RuntimeException("Enclosing Activity must implement JitsiMeetActivityInterface");
-        }
+    /**
+     * Called when the window containing this view gains or loses focus.
+     *
+     * @param hasFocus If the window of this view now has focus, {@code true};
+     * otherwise, {@code false}.
+     */
+    @Override
+    public void onWindowFocusChanged(boolean hasFocus) {
+        super.onWindowFocusChanged(hasFocus);
 
-        OngoingConferenceTracker.getInstance().addListener(this);
+        // https://github.com/mockingbot/react-native-immersive#restore-immersive-state
+        RNImmersiveModule immersive = RNImmersiveModule.getInstance();
+
+        if (hasFocus && immersive != null) {
+            immersive.emitImmersiveStateChangeEvent();
+        }
     }
 }
